pyproject.toml
[project]
name = "blog2kindle"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.12"
dependencies = [
    "beautifulsoup4>=4.14.3",
    "fastapi>=0.136.1",
    "html2text>=2025.4.15",
    "markdownify>=1.2.2",
    "mistune>=3.2.0",
    "pillow>=12.2.0",
    "pydantic-ai-slim[google,openai]>=1.87.0",
    "python-dotenv>=1.2.2",
    "python-multipart>=0.0.26",
    "requests>=2.33.1",
    "uvicorn>=0.46.0",
]
